move percy folder to test folder
