+++ .github/workflows/main.yml
@@ -45,7 +45,7 @@ jobs:
         run: npm run jest-screenshot
 
       - name: ▶️ Run Percy Screenshot
-        run: npm run test:percy
+        run: npm run percy-screenshot
         env:
           PERCY_TOKEN: ${{ secrets.PERCY_TOKEN }}
