+++ test/percy/global-teardown.js
