+++ test/percy/globals.js
