+++ test/percy/server/createConfig.js
@@ -7,7 +7,7 @@ import { buttonConfigs } from "../config";
 export const createConfig = () => {
   // eslint-disable-next-line no-sync
   fs.writeFileSync(
-    "./percy/files/buttonConfig.json",
+    "./test/percy/files/buttonConfig.json",
     JSON.stringify(buttonConfigs)
   );
 };
